poprawnienie pokrycia testow oraz dodanie flyway v2

+++ src/main/resources/db/migration/V1__init.sql
@@ -1,12 +1,14 @@
+-- src/main/resources/db/migration/V1__init.sql
+
 -- Tabela: roles
 CREATE TABLE roles (
-                       id SERIAL PRIMARY KEY,
+                       id BIGSERIAL PRIMARY KEY,
                        name VARCHAR(255) NOT NULL UNIQUE
 );
 
 -- Tabela: users
 CREATE TABLE users (
-                       id SERIAL PRIMARY KEY,
+                       id BIGSERIAL PRIMARY KEY,
                        username VARCHAR(255) NOT NULL UNIQUE,
                        password VARCHAR(255) NOT NULL,
                        email VARCHAR(255) NOT NULL UNIQUE
@@ -14,14 +16,14 @@ CREATE TABLE users (
 
 -- Tabela: user_roles (relacja many-to-many: users <-> roles)
 CREATE TABLE user_roles (
-                            user_id INTEGER NOT NULL REFERENCES users(id) ON DELETE CASCADE,
-                            role_id INTEGER NOT NULL REFERENCES roles(id) ON DELETE CASCADE,
+                            user_id BIGINT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+                            role_id BIGINT NOT NULL REFERENCES roles(id) ON DELETE CASCADE,
                             PRIMARY KEY (user_id, role_id)
 );
 
 -- Tabela: cars
 CREATE TABLE cars (
-                      id SERIAL PRIMARY KEY,
+                      id BIGSERIAL PRIMARY KEY,
                       make VARCHAR(100),
                       model VARCHAR(100),
                       year INT,
@@ -30,11 +32,10 @@ CREATE TABLE cars (
 
 -- Tabela: reservations
 CREATE TABLE reservations (
-                              id SERIAL PRIMARY KEY,
+                              id BIGSERIAL PRIMARY KEY,
                               start_date DATE NOT NULL,
                               end_date DATE NOT NULL,
-                              user_id INTEGER NOT NULL REFERENCES users(id) ON DELETE CASCADE,
-                              car_id INTEGER NOT NULL REFERENCES cars(id) ON DELETE CASCADE,
-    -- Możesz dodać inne pola jeśli masz w encji
+                              user_id BIGINT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+                              car_id BIGINT NOT NULL REFERENCES cars(id) ON DELETE CASCADE,
                               status VARCHAR(32)
 );
